requirements.txt
flask>=1.1
pandas==1.1.*
scipy==1.5.*
humanize
isodate
discord.py
tabulate
pdfkit
pydantic
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
